S13_Advanced_RAG/S13_requirements.txt
chromadb
pprint
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
python-dotenv
operator
langchain